测试PC 通过

+++ Assets/Scripts/HotFixMono/Game/GameMode.cs
@@ -10,7 +10,7 @@ namespace GameHotFix
     /// <summary>
     /// 游戏模式。
     /// </summary>
-    public enum GameMode : int
+    public enum GameMode : byte
     {
         /// <summary>
         /// 生存模式。
